requirements.txt
pefile
requests
reportlab
yara-python
python-magic
httpx
python-dotenv
